% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/113.ConfidenceIntervals_ADJ_n_x.R
\name{ciAAllx}
\alias{ciAAllx}
\title{CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) given x and n}
\usage{
ciAAllx(x, n, alp, h)
}
\arguments{
\item{x}{- Number of success}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\value{
A dataframe with
 \item{name }{- Name of the method}
 \item{x }{- Number of successes (positive samples)}
 \item{LLT }{ - Lower limit}
 \item{ULT }{ - Upper Limit}
 \item{LABB }{ - Lower Abberation}
 \item{UABB }{ - Upper Abberation}
 \item{ZWI }{ - Zero Width Interval}
}
\description{
CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) given x and n
}
\details{
The Confidence Interval of using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) for \code{n} given \code{alp}, \code{x} and \code{h}
}
\examples{
x=5; n=5; alp=0.05;h=2
ciAAllx(x,n,alp,h)
}
\references{
[1] 1998 Agresti A and Coull BA.
Approximate is better than "Exact" for interval estimation of binomial proportions.
The American Statistician: 52; 119 - 126.

[2] 1998 Newcombe RG.
Two-sided confidence intervals for the single proportion: Comparison of seven methods.
Statistics in Medicine: 17; 857 - 872.

[3] 2008 Pires, A.M., Amado, C.
Interval Estimators for a Binomial Proportion: Comparison of Twenty Methods.
REVSTAT - Statistical Journal, 6, 165-197.
}
\seealso{
\code{\link{prop.test} and \link{binom.test}} for equivalent base Stats R functionality,
   \code{\link[binom]{binom.confint}}  provides similar functionality for 11 methods,
   \code{\link[PropCIs]{wald2ci}} which provides multiple functions for CI calculation ,
   \code{\link[BlakerCI]{binom.blaker.limits}} which calculates Blaker CI which is not covered here and
   \code{\link[prevalence]{propCI}} which provides similar functionality.

Other Adjusted methods of CI estimation  given x & n: 
\code{\link{PlotciAAllx}()},
\code{\link{ciAASx}()},
\code{\link{ciALRx}()},
\code{\link{ciALTx}()},
\code{\link{ciASCx}()},
\code{\link{ciATWx}()},
\code{\link{ciAWDx}()}
}
\concept{Adjusted methods of CI estimation  given x & n}
