% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/113.ConfidenceIntervals_ADJ_n_x.R
\name{ciAASx}
\alias{ciAASx}
\title{Adjusted ArcSine method of CI estimation}
\usage{
ciAASx(x, n, alp, h)
}
\arguments{
\item{x}{- Number of successes}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\value{
A dataframe with
 \item{x}{  Number of successes (positive samples)}
 \item{LAASx }{   ArcSine Lower limit}
 \item{UAASx }{   ArcSine Upper Limit}
 \item{LABB }{   ArcSine Lower Abberation}
 \item{UABB }{   ArcSine Upper Abberation}
 \item{ZWI }{   Zero Width Interval}
}
\description{
Adjusted ArcSine method of CI estimation
}
\details{
Wald-type interval for the arcsine transformation of the parameter \code{p}
for the modified data \eqn{x + h} and \eqn{n + (2*h)} , where \eqn{h > 0} and for
the given \code{x} and \code{n}.
}
\examples{
x=5; n=5; alp=0.05;h=2
ciAASx(x,n,alp,h)
}
\references{
[1] 1998 Agresti A and Coull BA.
Approximate is better than "Exact" for interval estimation of binomial proportions.
The American Statistician: 52; 119 - 126.

[2] 1998 Newcombe RG.
Two-sided confidence intervals for the single proportion: Comparison of seven methods.
Statistics in Medicine: 17; 857 - 872.

[3] 2008 Pires, A.M., Amado, C.
Interval Estimators for a Binomial Proportion: Comparison of Twenty Methods.
REVSTAT - Statistical Journal, 6, 165-197.
}
\seealso{
\code{\link{prop.test} and \link{binom.test}} for equivalent base Stats R functionality,
   \code{\link[binom]{binom.confint}}  provides similar functionality for 11 methods,
   \code{\link[PropCIs]{wald2ci}} which provides multiple functions for CI calculation ,
   \code{\link[BlakerCI]{binom.blaker.limits}} which calculates Blaker CI which is not covered here and
   \code{\link[prevalence]{propCI}} which provides similar functionality.

Other Adjusted methods of CI estimation  given x & n: \code{\link{ciALRx}},
  \code{\link{ciALTx}}, \code{\link{ciATWx}},
  \code{\link{ciAWDx}}
}

