% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/123.ConfidenceIntervals_CC_n_x.R
\name{ciCASx}
\alias{ciCASx}
\title{Continuity corrected ArcSine method of CI estimation}
\usage{
ciCASx(x, n, alp, c)
}
\arguments{
\item{x}{- Number of successes}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{c}{- Continuity correction}
}
\value{
A dataframe with 
 \item{x}{  Number of successes (positive samples)}
 \item{LCAx }{   ArcSine Lower limit}
 \item{UCAx }{   ArcSine Upper Limit}
 \item{LABB }{   ArcSine Lower Abberation}
 \item{UABB }{   ArcSine Upper Abberation}
 \item{ZWI }{   Zero Width Interval}
}
\description{
Continuity corrected ArcSine method of CI estimation
}
\details{
Wald-type interval for the arcsine transformation using the test statistic
\eqn{\frac{\mid{sin^{-1}\hat{p}-sin^{-1}p \mid-c}{SE}}}{(abs(sin^(-1)phat-sin^(-1)p)-c)/SE}
 where \eqn{c > 0} is a constant for continuity correction and for all \eqn{x = 0, 1, 2 ..n}
}
\examples{
x=5; n=5; alp=0.05;c=1/2*n
ciCASx(x,n,alp,c)
}
\references{
[1] 1998 Agresti A and Coull BA. 
Approximate is better than "Exact" for interval estimation of binomial proportions. 
The American Statistician: 52; 119 - 126.

[2] 1998 Newcombe RG. 
Two-sided confidence intervals for the single proportion: Comparison of seven methods. 
Statistics in Medicine: 17; 857 - 872.

[3] 2008 Pires, A.M., Amado, C. 
Interval Estimators for a Binomial Proportion: Comparison of Twenty Methods. 
REVSTAT - Statistical Journal, 6, 165-197.
}
\seealso{
\code{\link{prop.test} and \link{binom.test}} for equivalent base Stats R functionality,
   \code{\link[=binom]{binom.confint}}  provides similar functionality for 11 methods,
   \code{\link[=PropCIs]{wald2ci}} which provides multiple functions for CI calculation ,
   \code{\link[=BlakerCI]{binom.blaker.limits}} which calculates Blaker CI which is not covered here and
   \code{\link[=prevalence]{propCI}} which provides similar functionality.
}

