% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/511.Error-Failure_LimitBased_ADJ_All.R
\name{errALR}
\alias{errALR}
\title{Calculates error, long term power and pass/fail criteria for adjusted Likelihood Ratio method}
\usage{
errALR(n, alp, h, phi, f)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}

\item{phi}{- Null hypothesis value}

\item{f}{- Failure criterion}
}
\value{
A dataframe with
 \item{delalp}{ Delta-alpha is the increase of the nominal error with respect to real error}
 \item{theta}{ Long term power of the test}
 \item{Fail_Pass}{Fail/pass based on the input f criterion}
}
\description{
Calculates error, long term power and pass/fail criteria for adjusted Likelihood Ratio method
}
\details{
Evaluation of adjusted Likelihood ratio limits using error due to the
difference of achieved and nominal level of significance for the \eqn{n + 1} intervals
}
\examples{
n=20; alp=0.05; h=2;phi=0.99; f=-2
errALR(n,alp,h,phi,f)
}
\references{
[1] 2014 Martín Andrés, A. and Álvarez Hernández, M.
Two-tailed asymptotic inferences for a proportion.
Journal of Applied Statistics, 41, 7, 1516-1529
}
\seealso{
Other Error for adjusted methods: \code{\link{PloterrAAS}},
  \code{\link{PloterrAAll}}, \code{\link{PloterrALR}},
  \code{\link{PloterrALT}}, \code{\link{PloterrASC}},
  \code{\link{PloterrATW}}, \code{\link{PloterrAWD}},
  \code{\link{errAAS}}, \code{\link{errAAll}},
  \code{\link{errALT}}, \code{\link{errASC}},
  \code{\link{errATW}}, \code{\link{errAWD}}
}

