% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/521.Error-Failure_Limitbased_CC_All.R
\name{errCAS}
\alias{errCAS}
\title{Calculates error, long term power and pass/fail criteria for continuity corrected ArcSine method}
\usage{
errCAS(n, alp, phi, c, f)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{phi}{- Null hypothesis value}

\item{c}{- Continuity correction}

\item{f}{- Failure criterion}
}
\value{
A dataframe with 
 \item{delalp}{ Delta-alpha is the increase of the nominal error with respect to real error}
 \item{theta}{ Long term power of the test}
 \item{Fail_Pass}{Fail/pass based on the input f criterion}
}
\description{
Calculates error, long term power and pass/fail criteria for continuity corrected ArcSine method
}
\details{
Evaluation of continuity corrected Wald-type interval for the arcsine 
transformation of the parameter \code{p} 
using error due to the difference of achieved and nominal level of significance for the \eqn{n + 1} intervals
}
\examples{
n=5; alp=0.05; phi=0.05;c=1/(2*n); f=-2
errCAS(n,alp,phi,c,f)
}
\references{
[1] 2014 Martín Andrés, A. and Álvarez Hernández, M. 
Two-tailed asymptotic inferences for a proportion. 
Journal of Applied Statistics, 41, 7, 1516-1529
}
\seealso{
Other Error for continuity corrected methods: \code{\link{PloterrCAS}},
  \code{\link{PloterrCAll}}, \code{\link{PloterrCLT}},
  \code{\link{PloterrCSC}}, \code{\link{PloterrCTW}},
  \code{\link{PloterrCWD}}, \code{\link{errCAll}},
  \code{\link{errCLT}}, \code{\link{errCSC}},
  \code{\link{errCTW}}, \code{\link{errCWD}}
}

