% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/501.Error-Failure_LimitBased_BASE_All.R
\name{errEX}
\alias{errEX}
\title{Calculates error, long term power and pass/fail criteria for Exact method}
\usage{
errEX(n, alp, phi, f, e)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{phi}{- Null hypothesis value}

\item{f}{- Failure criterion}

\item{e}{- Exact method indicator  in [0, 1] {1: Clopper Pearson, 0.5: Mid P}
The input can also be a range of values between 0 and 1.}
}
\value{
A dataframe with
 \item{delalp}{ Delta-alpha is the increase of the nominal error with respect to real error}
 \item{theta}{ Long term power of the test}
 \item{Fail_Pass}{Fail/pass based on the input f criterion}
}
\description{
Calculates error, long term power and pass/fail criteria for Exact method
}
\details{
Evaluation of Confidence interval for \code{p}
based on inverting equal-tailed binomial tests with null hypothesis \eqn{H0: p = p0}
using error due to the difference of achieved and nominal level of significance for the \eqn{n + 1} intervals
}
\examples{
n=20; alp=0.05;phi=0.05; f=-2;e=0.5 # Mid-p
errEX(n,alp,phi,f,e)
n=20; alp=0.05;phi=0.05; f=-2;e=1 #Clopper-Pearson
errEX(n,alp,phi,f,e)
n=20; alp=0.05;phi=0.05; f=-2;e=c(0.1,0.5,0.95,1) #Range including Mid-p and Clopper-Pearson
errEX(n,alp,phi,f,e)
}
\references{
[1] 2014 Martin Andres, A. and Alvarez Hernandez, M.
Two-tailed asymptotic inferences for a proportion.
Journal of Applied Statistics, 41, 7, 1516-1529
}
\seealso{
Other Error for base methods: 
\code{\link{PloterrAS}()},
\code{\link{PloterrAll}()},
\code{\link{PloterrBA}()},
\code{\link{PloterrEX}()},
\code{\link{PloterrLR}()},
\code{\link{PloterrLT}()},
\code{\link{PloterrSC}()},
\code{\link{PloterrTW}()},
\code{\link{PloterrWD}()},
\code{\link{errAS}()},
\code{\link{errAll}()},
\code{\link{errBA}()},
\code{\link{errLR}()},
\code{\link{errLT}()},
\code{\link{errSC}()},
\code{\link{errTW}()},
\code{\link{errWD}()}
}
\concept{Error for base methods}
