% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/631.BayesFactors.R
\name{hypotestBAF6}
\alias{hypotestBAF6}
\title{Bayesain Hypothesis testing : Hypothesis 6: Theta < Theta1 Vs Theta > Theta2}
\usage{
hypotestBAF6(n, th1, a1, b1, th2, a2, b2)
}
\arguments{
\item{n}{- Number of trials from data}

\item{th1}{- Hypothetical parameter for H1}

\item{a1}{- Priors for hypothesis H1}

\item{b1}{- Priors for hypothesis H1}

\item{th2}{- Hypothetical parameter for H2}

\item{a2}{- Priors for hypothesis H2}

\item{b2}{- Priors for hypothesis H2}
}
\value{
A dataframe with 
 \item{x}{ Number of successes}
 \item{BaFa12}{ Bayesian Factor}
}
\description{
Bayesain Hypothesis testing : Hypothesis 6: Theta < Theta1 Vs Theta > Theta2
}
\details{
Computes Bayes factor under Beta-Binomial model for the 
model: \eqn{p < p1} Vs \eqn{p > p2} from the given number of trials \code{n} and for all number 
of successes \eqn{x = 0, 1, 2......n }
}
\examples{
n=1;th1=0.5; a1=1; b1=1; th2=0.9; a2=0.5; b2=0.5
hypotestBAF6(n,th1,a1,b1,th2,a2,b2)
}
\references{
[1] 2006  Ghosh M, Delampady M and Samanta T. 
An introduction to Bayesian analysis: Theory and Methods. 
Springer, New York
[2] 2014 Sakthivel S, Subbiah M and Ramakrishnan R 
Default prior approach for Bayesian testing of hypotheses involving single binomial proportion 
International Journal of Statistics and Analysis, 4 (2), 139 - 153
}
\seealso{
Other Hypothesis testing: \code{\link{hypotestBAF1x}},
  \code{\link{hypotestBAF1}}, \code{\link{hypotestBAF2x}},
  \code{\link{hypotestBAF2}}, \code{\link{hypotestBAF3x}},
  \code{\link{hypotestBAF3}}, \code{\link{hypotestBAF4x}},
  \code{\link{hypotestBAF4}}, \code{\link{hypotestBAF5x}},
  \code{\link{hypotestBAF5}}, \code{\link{hypotestBAF6x}}
}

