% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/324.Expec_Leng_GENERAL_SIMULATEDp.R
\name{lengthGEN}
\alias{lengthGEN}
\title{Calculates the sum of lengths for a specific LL and UL}
\usage{
lengthGEN(n, LL, UL)
}
\arguments{
\item{n}{- Number of trials}

\item{LL}{- Lower limit}

\item{UL}{- Upper limit}
}
\value{
A number
 \item{sumLE}{  Sum of the lengths}
}
\description{
Calculates the sum of lengths for a specific LL and UL
}
\details{
Evaluation of intervals obtained from any method using sum
of the lengths for the \eqn{n + 1} intervals
}
\examples{
LL=c(0,0.01,0.0734,0.18237,0.3344,0.5492)		#Lower and Upper Limits
UL=c(0.4507,0.6655,0.8176,0.9265,0.9899,1)
n= 5;
lengthGEN(n,LL,UL)
}
\seealso{
Other Expected length: \code{\link{PlotexplGEN}},
  \code{\link{PlotexplSIM}}
}

