% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/401.p-Confidence_p-Bias_BASE_All.R
\name{pCOpBITW}
\alias{pCOpBITW}
\title{T-Wald method for p-confidence and p-bias for a given n and alpha level}
\usage{
pCOpBITW(n, alp)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}
}
\value{
A dataframe with
 \item{x1}{  Number of successes (positive samples)}
 \item{pconf }{   p-Confidence}
 \item{pbias }{   p-Bias}
}
\description{
T-Wald method for p-confidence and p-bias for a given n and alpha level
}
\details{
Evaluation of approximate method based on a t_approximation of the standardized point estimator using p-confidence and p-bias for the \eqn{n + 1} intervals
}
\examples{
n=5; alp=0.05
pCOpBITW(n,alp)
}
\references{
[1] 2005 Vos PW and Hudson S.
Evaluation Criteria for Discrete Confidence Intervals: Beyond Coverage and Length.
The American Statistician: 59; 137 - 142.
}
\seealso{
Other p-confidence and p-bias of base methods: \code{\link{PlotpCOpBIAS}},
  \code{\link{PlotpCOpBIAll}}, \code{\link{PlotpCOpBIBA}},
  \code{\link{PlotpCOpBIEX}}, \code{\link{PlotpCOpBILR}},
  \code{\link{PlotpCOpBILT}}, \code{\link{PlotpCOpBISC}},
  \code{\link{PlotpCOpBITW}}, \code{\link{PlotpCOpBIWD}},
  \code{\link{pCOpBIAS}}, \code{\link{pCOpBIAll}},
  \code{\link{pCOpBIBA}}, \code{\link{pCOpBIEX}},
  \code{\link{pCOpBILR}}, \code{\link{pCOpBILT}},
  \code{\link{pCOpBISC}}, \code{\link{pCOpBIWD}}
}

