% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/112.ConfidenceIntervals_ADJ_n_Graph.R
\name{PlotciAAll}
\alias{PlotciAAll}
\title{Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)}
\usage{
PlotciAAll(n, alp, h)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\description{
Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\details{
The plots of the Confidence Interval using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) for \code{n} given \code{alp} and \code{h}
}
\examples{
n=5; alp=0.05;h=2
PlotciAAll(n,alp,h)
}
\seealso{
Other Adjusted methods of CI estimation: \code{\link{PlotciAAS}},
  \code{\link{PlotciAAllg}}, \code{\link{PlotciALR}},
  \code{\link{PlotciALT}}, \code{\link{PlotciASC}},
  \code{\link{PlotciATW}}, \code{\link{PlotciAWD}},
  \code{\link{ciAAS}}, \code{\link{ciAAll}},
  \code{\link{ciALR}}, \code{\link{ciALT}},
  \code{\link{ciASC}}, \code{\link{ciATW}},
  \code{\link{ciAWD}}
}

