% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/112.ConfidenceIntervals_ADJ_n_Graph.R
\name{PlotciAAllg}
\alias{PlotciAAllg}
\title{Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) grouped by x value}
\usage{
PlotciAAllg(n, alp, h)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- adjustment}
}
\description{
Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) grouped by x value
}
\details{
The plot of Confidence Interval of \code{n} given \code{alp} and \code{h} grouped by x for 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\examples{
n=5; alp=0.05; h=2
PlotciAAllg(n,alp,h)
}
\seealso{
Other Adjusted methods of CI estimation: 
\code{\link{PlotciAAS}()},
\code{\link{PlotciAAll}()},
\code{\link{PlotciALR}()},
\code{\link{PlotciALT}()},
\code{\link{PlotciASC}()},
\code{\link{PlotciATW}()},
\code{\link{PlotciAWD}()},
\code{\link{ciAAS}()},
\code{\link{ciAAll}()},
\code{\link{ciALR}()},
\code{\link{ciALT}()},
\code{\link{ciASC}()},
\code{\link{ciATW}()},
\code{\link{ciAWD}()}
}
\concept{Adjusted methods of CI estimation}
