% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/113.ConfidenceIntervals_ADJ_n_x.R
\name{PlotciAAllxg}
\alias{PlotciAAllxg}
\title{Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) given x & n grouped by x value}
\usage{
PlotciAAllxg(x, n, alp, h)
}
\arguments{
\item{x}{- Number of success}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\description{
Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) given x & n grouped by x value
}
\details{
The plots of confidence intervals of using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) grouped by x for \code{n} given \code{alp}, \code{x} and \code{h}
}
\examples{
x=5; n=5; alp=0.5;h=2
PlotciAAllxg(x,n,alp,h)
}
\seealso{
\code{\link{prop.test} and \link{binom.test}} for equivalent base Stats R functionality,
   \code{\link[=binom]{binom.confint}}  provides similar functionality for 11 methods,
   \code{\link[=PropCIs]{wald2ci}} which provides multiple functions for CI calculation ,
   \code{\link[=BlakerCI]{binom.blaker.limits}} which calculates Blaker CI which is not covered here and
   \code{\link[=prevalence]{propCI}} which provides similar functionality.
}

