% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/102.Confidence_base_n_Graph.R
\name{PlotciAll}
\alias{PlotciAll}
\title{Plots the CI estimation of 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) methods}
\usage{
PlotciAll(n, alp)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}
}
\description{
Plots the CI estimation of 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) methods
}
\details{
The plot of Confidence Interval of 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) for \code{n} given \code{alp}.
}
\examples{
n=5; alp=0.05;
PlotciAll(n,alp)
}
\seealso{
Other Basic methods of CI estimation: \code{\link{PlotciAS}},
  \code{\link{PlotciAllg}}, \code{\link{PlotciBA}},
  \code{\link{PlotciEX}}, \code{\link{PlotciLR}},
  \code{\link{PlotciLT}}, \code{\link{PlotciSC}},
  \code{\link{PlotciTW}}, \code{\link{PlotciWD}},
  \code{\link{ciAS}}, \code{\link{ciAll}},
  \code{\link{ciBA}}, \code{\link{ciEX}},
  \code{\link{ciLR}}, \code{\link{ciLT}},
  \code{\link{ciSC}}, \code{\link{ciTW}},
  \code{\link{ciWD}}
}

