% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/112.ConfidenceIntervals_ADJ_n_Graph.R
\name{PlotciATW}
\alias{PlotciATW}
\title{Plot of CI estimation of adjusted Wald-T given n, alp and h}
\usage{
PlotciATW(n, alp, h)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\description{
Plot of CI estimation of adjusted Wald-T given n, alp and h
}
\details{
The plots of the Confidence Interval using adjusted Wald-T method for \code{n} given \code{alp} and \code{h}
}
\examples{
n=5; alp=0.05;h=2
PlotciATW(n,alp,h)
}
\seealso{
Other Adjusted methods of CI estimation: \code{\link{PlotciAAS}},
  \code{\link{PlotciAAllg}}, \code{\link{PlotciAAll}},
  \code{\link{PlotciALR}}, \code{\link{PlotciALT}},
  \code{\link{PlotciASC}}, \code{\link{PlotciAWD}},
  \code{\link{ciAAS}}, \code{\link{ciAAll}},
  \code{\link{ciALR}}, \code{\link{ciALT}},
  \code{\link{ciASC}}, \code{\link{ciATW}},
  \code{\link{ciAWD}}
}

