% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/122.ConfidenceIntervals_CC_n_Graph.R
\name{PlotciCAllg}
\alias{PlotciCAllg}
\title{Plots the CI estimation of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) grouped by x value}
\usage{
PlotciCAllg(n, alp, c)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{c}{- Continuity correction}
}
\description{
Plots the CI estimation of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) grouped by x value
}
\details{
Plots the Confidence Interval for 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) grouped by x for \code{n} given \code{alp} along with Continuity correction \code{c}
}
\examples{
n=5; alp=0.05; c=1/(2*n)
PlotciCAllg(n,alp,c)
}
\seealso{
Other Continuity correction methods of CI estimation: 
\code{\link{PlotciCAS}()},
\code{\link{PlotciCAll}()},
\code{\link{PlotciCLT}()},
\code{\link{PlotciCSC}()},
\code{\link{PlotciCTW}()},
\code{\link{PlotciCWD}()},
\code{\link{ciCAS}()},
\code{\link{ciCAll}()},
\code{\link{ciCLT}()},
\code{\link{ciCSC}()},
\code{\link{ciCTW}()},
\code{\link{ciCWD}()}
}
\concept{Continuity correction methods of CI estimation}
