% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/123.ConfidenceIntervals_CC_n_x.R
\name{PlotciCAllx}
\alias{PlotciCAllx}
\title{Plots the CI estimation of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) given n, alp and x with continuity correction c}
\usage{
PlotciCAllx(x, n, alp, c)
}
\arguments{
\item{x}{- Number of sucess}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{c}{- Continuity correction}
}
\description{
Plots the CI estimation of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) given n, alp and x with continuity correction c
}
\details{
Plots the Confidence Interval of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) for \code{n} given \code{alp}, \code{x} and and \code{c}
}
\examples{
x=5; n=5; alp=0.05;c=1/(2*n)
PlotciCAllx(x,n,alp,c)
}
\seealso{
Other Continuity correction methods of CI estimation given x and n: 
\code{\link{PlotciCAllxg}()},
\code{\link{ciCAllx}()},
\code{\link{ciCLTx}()},
\code{\link{ciCSCx}()},
\code{\link{ciCTWx}()},
\code{\link{ciCWDx}()}
}
\concept{Continuity correction methods of CI estimation given x and n}
