% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/123.ConfidenceIntervals_CC_n_x.R
\name{PlotciCAllxg}
\alias{PlotciCAllxg}
\title{Plots the CI estimation of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) given x & n grouped by x value}
\usage{
PlotciCAllxg(x, n, alp, c)
}
\arguments{
\item{x}{- Number of sucess}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{c}{- Continuity correction}
}
\description{
Plots the CI estimation of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) given x & n grouped by x value
}
\details{
Plots he Confidence Interval of 5 continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine) for \code{n} given \code{alp}, \code{c} and \code{x} grouped by x
}
\examples{
x=5; n=5; alp=0.05;c=1/(2*n)
PlotciCAllxg(x,n,alp,c)
}
\seealso{
Other Continuity correction methods of CI estimation given x and n: 
\code{\link{PlotciCAllx}()},
\code{\link{ciCAllx}()},
\code{\link{ciCLTx}()},
\code{\link{ciCSCx}()},
\code{\link{ciCTWx}()},
\code{\link{ciCWDx}()}
}
\concept{Continuity correction methods of CI estimation given x and n}
