% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/104.ConfidenceIntervals_BASE_n_x_Graph.R
\name{PlotciEXx}
\alias{PlotciEXx}
\title{Plots the CI estimation of the exact method given n, alp and x}
\usage{
PlotciEXx(x, n, alp, e)
}
\arguments{
\item{x}{- Number of sucess}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{e}{- Exact method indicator  in [0, 1] {1: Clopper Pearson, 0.5: Mid P}}
}
\description{
Plots the CI estimation of the exact method given n, alp and x
}
\details{
Plot of the Confidence interval for exact method
}
\examples{
x=5; n=5; alp=0.05;e=0.5
PlotciEXx(x,n,alp,e) #Mid-p
x=5; n=5; alp=0.05;e=1 #Clopper Pearson
PlotciEXx(x,n,alp,e) 
x=5; n=5; alp=0.05;e=c(0.1,0.5,0.95,1) #Range including Mid-p and Clopper-Pearson
PlotciEXx(x,n,alp,e)
}
\seealso{
Other Base methods of CI estimation given x & n: \code{\link{PlotciAllxg}},
  \code{\link{PlotciAllx}}, \code{\link{ciASx}},
  \code{\link{ciAllx}}, \code{\link{ciBAx}},
  \code{\link{ciEXx}}, \code{\link{ciLRx}},
  \code{\link{ciLTx}}, \code{\link{ciSCx}},
  \code{\link{ciTWx}}, \code{\link{ciWDx}}
}

