% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/102.Confidence_base_n_Graph.R
\name{PlotciLR}
\alias{PlotciLR}
\title{Plots the CI estimation of Likelihood Ratio method}
\usage{
PlotciLR(n, alp)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}
}
\description{
Plots the CI estimation of Likelihood Ratio method
}
\details{
The plot of Confidence Interval of \code{n} given \code{alp} using Likelihood Ratio method
}
\examples{
n=5; alp=0.05
PlotciLR(n,alp)
}
\seealso{
Other Basic methods of CI estimation: \code{\link{PlotciAS}},
  \code{\link{PlotciAllg}}, \code{\link{PlotciAll}},
  \code{\link{PlotciBA}}, \code{\link{PlotciEX}},
  \code{\link{PlotciLT}}, \code{\link{PlotciSC}},
  \code{\link{PlotciTW}}, \code{\link{PlotciWD}},
  \code{\link{ciAS}}, \code{\link{ciAll}},
  \code{\link{ciBA}}, \code{\link{ciEX}},
  \code{\link{ciLR}}, \code{\link{ciLT}},
  \code{\link{ciSC}}, \code{\link{ciTW}},
  \code{\link{ciWD}}
}

