% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/102.Confidence_base_n_Graph.R
\name{PlotciSC}
\alias{PlotciSC}
\title{Plots the CI estimation of Score method}
\usage{
PlotciSC(n, alp)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}
}
\description{
Plots the CI estimation of Score method
}
\details{
The plot of Confidence Interval of \code{n} given \code{alp} using Score method
}
\examples{
n=5; alp=0.05
PlotciSC(n,alp)
}
\seealso{
Other Basic methods of CI estimation: \code{\link{PlotciAS}},
  \code{\link{PlotciAllg}}, \code{\link{PlotciAll}},
  \code{\link{PlotciBA}}, \code{\link{PlotciEX}},
  \code{\link{PlotciLR}}, \code{\link{PlotciLT}},
  \code{\link{PlotciTW}}, \code{\link{PlotciWD}},
  \code{\link{ciAS}}, \code{\link{ciAll}},
  \code{\link{ciBA}}, \code{\link{ciEX}},
  \code{\link{ciLR}}, \code{\link{ciLT}},
  \code{\link{ciSC}}, \code{\link{ciTW}},
  \code{\link{ciWD}}
}

