% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/202.CoverageProb_BASE_All_Graph.R
\name{PlotcovpAS}
\alias{PlotcovpAS}
\title{Plots Coverage Probability for base ArcSine method}
\usage{
PlotcovpAS(n, alp, a, b, t1, t2)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}

\item{t1}{- Lower tolerance limit to check the spread of coverage Probability}

\item{t2}{- Upper tolerance limit to check the spread of coverage Probability}
}
\description{
Plots Coverage Probability for base ArcSine method
}
\details{
Plots Coverage Probability for base ArcSine method
}
\examples{
n= 10; alp=0.05; a=1; b=1; t1=0.93; t2=0.97
PlotcovpAS(n,alp,a,b,t1,t2)
}
\seealso{
Other Basic coverage probability methods: \code{\link{PlotcovpAll}},
  \code{\link{PlotcovpBA}}, \code{\link{PlotcovpEX}},
  \code{\link{PlotcovpLR}}, \code{\link{PlotcovpLT}},
  \code{\link{PlotcovpSC}}, \code{\link{PlotcovpTW}},
  \code{\link{PlotcovpWD}}, \code{\link{covpAS}},
  \code{\link{covpAll}}, \code{\link{covpBA}},
  \code{\link{covpEX}}, \code{\link{covpLR}},
  \code{\link{covpLT}}, \code{\link{covpSC}},
  \code{\link{covpTW}}, \code{\link{covpWD}}
}

