% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/202.CoverageProb_BASE_All_Graph.R
\name{PlotcovpBA}
\alias{PlotcovpBA}
\title{Graphs of  Coverage Probability of the Bayesian method}
\usage{
PlotcovpBA(n, alp, a, b, t1, t2, a1, a2)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}

\item{t1}{- Lower tolerance limit to check the spread of coverage Probability}

\item{t2}{- Upper tolerance limit to check the spread of coverage Probability}

\item{a1}{- Beta Prior Parameters for Bayesian estimation}

\item{a2}{- Beta Prior Parameters for Bayesian estimation}
}
\description{
Graphs of  Coverage Probability of the Bayesian method
}
\details{
The  graphs of  Coverage Probability of Bayesian method
}
\examples{
n= 10; alp=0.05; a=1;b=1; t1=0.93;t2=0.97;a1=1;a2=1
PlotcovpBA(n,alp,a,b,t1,t2,a1,a2)
}
\seealso{
Other Basic coverage probability methods: \code{\link{PlotcovpAS}},
  \code{\link{PlotcovpAll}}, \code{\link{PlotcovpEX}},
  \code{\link{PlotcovpLR}}, \code{\link{PlotcovpLT}},
  \code{\link{PlotcovpSC}}, \code{\link{PlotcovpTW}},
  \code{\link{PlotcovpWD}}, \code{\link{covpAS}},
  \code{\link{covpAll}}, \code{\link{covpBA}},
  \code{\link{covpEX}}, \code{\link{covpLR}},
  \code{\link{covpLT}}, \code{\link{covpSC}},
  \code{\link{covpTW}}, \code{\link{covpWD}}
}

