% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/222.CoverageProb_CC_All_Graph.R
\name{PlotcovpCTW}
\alias{PlotcovpCTW}
\title{Graphs  of Coverage Probability for continuity corrected Wald-T method}
\usage{
PlotcovpCTW(n, alp, c, a, b, t1, t2)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{c}{- Continiuty correction}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}

\item{t1}{- Lower tolerance limit to check the spread of coverage Probability}

\item{t2}{- Upper tolerance limit to check the spread of coverage Probability}
}
\description{
Graphs  of Coverage Probability for continuity corrected Wald-T method
}
\details{
The plot of the  Coverage Probability of continuity corrected Wald-T method
for \code{n} given \code{alp}, \code{h}, \code{a}, \code{b}, \code{t1} and  \code{t2} using all the methods
}
\examples{
n= 10; alp=0.05; c=1/(2*n);a=1;b=1; t1=0.93;t2=0.97
PlotcovpCTW(n,alp,c,a,b,t1,t2)
}
\seealso{
Other Coverage probability for continuity corrected methods: \code{\link{PlotcovpCAS}},
  \code{\link{PlotcovpCAll}}, \code{\link{PlotcovpCLT}},
  \code{\link{PlotcovpCSC}}, \code{\link{PlotcovpCWD}},
  \code{\link{covpCAS}}, \code{\link{covpCAll}},
  \code{\link{covpCLT}}, \code{\link{covpCSC}},
  \code{\link{covpCTW}}, \code{\link{covpCWD}}
}

