% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/202.CoverageProb_BASE_All_Graph.R
\name{PlotcovpEX}
\alias{PlotcovpEX}
\title{Graphs of Coverage Probability - exact method}
\usage{
PlotcovpEX(n, alp, e, a, b, t1, t2)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{e}{-  Exact method indicator (1:Clop-Pear,0.5:MID-p)}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}

\item{t1}{- Lower tolerance limit to check the spread of coverage Probability}

\item{t2}{- Upper tolerance limit to check the spread of coverage Probability}
}
\description{
Graphs of Coverage Probability - exact method
}
\details{
The  graphs of basic Coverage Probability methods
}
\examples{
\dontrun{
n= 10; alp=0.05; e=0.5; a=1;b=1; t1=0.93;t2=0.97 # Mid-p
PlotcovpEX(n,alp,e,a,b,t1,t2)
n= 10; alp=0.05; e=1; a=1;b=1; t1=0.93;t2=0.97 #Clop-Pear
PlotcovpEX(n,alp,e,a,b,t1,t2)
n=5; alp=0.05;
e=c(0.1,0.5,0.95,1) #Range including Mid-p and Clopper-Pearson
a=1;b=1; t1=0.93;t2=0.97
PlotcovpEX(n,alp,e,a,b,t1,t2)
}
}
\seealso{
Other Basic coverage probability methods: \code{\link{PlotcovpAS}},
  \code{\link{PlotcovpAll}}, \code{\link{PlotcovpBA}},
  \code{\link{PlotcovpLR}}, \code{\link{PlotcovpLT}},
  \code{\link{PlotcovpSC}}, \code{\link{PlotcovpTW}},
  \code{\link{PlotcovpWD}}, \code{\link{covpAS}},
  \code{\link{covpAll}}, \code{\link{covpBA}},
  \code{\link{covpEX}}, \code{\link{covpLR}},
  \code{\link{covpLT}}, \code{\link{covpSC}},
  \code{\link{covpTW}}, \code{\link{covpWD}}
}

