% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/511.Error-Failure_LimitBased_ADJ_All.R, R/512.Error-Failure_LimitBased_ADJ_All_Graph.R
\name{PloterrAAll}
\alias{PloterrAAll}
\title{Plots error, long term power and pass/fail criteria using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)}
\usage{
PloterrAAll(n, alp, h, phi, f)

PloterrAAll(n, alp, h, phi, f)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}

\item{phi}{- Null hypothesis value}

\item{f}{- Failure criterion}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}

\item{phi}{- Null hypothesis value}

\item{f}{- Failure criterion}
}
\description{
Plots error, long term power and pass/fail criteria using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)

Plots error, long term power and pass/fail criteria using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\details{
Plots of error, long term power and pass/fail
criteria using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)

Calculates  error, long term power and pass/fail
criteria using 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\examples{
n=20; alp=0.05;h=2; phi=0.99; f=-2
PloterrAAll(n,alp,h,phi,f)
n=20; alp=0.05;h=2; phi=0.99; f=-2
PloterrAAll(n,alp,h,phi,f)
}
\seealso{
Other Error for adjusted methods: \code{\link{PloterrAAS}},
  \code{\link{PloterrALR}}, \code{\link{PloterrALT}},
  \code{\link{PloterrASC}}, \code{\link{PloterrATW}},
  \code{\link{PloterrAWD}}, \code{\link{errAAS}},
  \code{\link{errAAll}}, \code{\link{errALR}},
  \code{\link{errALT}}, \code{\link{errASC}},
  \code{\link{errATW}}, \code{\link{errAWD}}

Other Error for adjusted methods: \code{\link{PloterrAAS}},
  \code{\link{PloterrALR}}, \code{\link{PloterrALT}},
  \code{\link{PloterrASC}}, \code{\link{PloterrATW}},
  \code{\link{PloterrAWD}}, \code{\link{errAAS}},
  \code{\link{errAAll}}, \code{\link{errALR}},
  \code{\link{errALT}}, \code{\link{errASC}},
  \code{\link{errATW}}, \code{\link{errAWD}}
}

