% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/512.Error-Failure_LimitBased_ADJ_All_Graph.R
\name{PloterrAWD}
\alias{PloterrAWD}
\title{Plots error, long term power and pass/fail criteria using adjusted Wald method}
\usage{
PloterrAWD(n, alp, h, phi, f)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}

\item{phi}{- Null hypothesis value}

\item{f}{- Failure criterion}
}
\description{
Plots error, long term power and pass/fail criteria using adjusted Wald method
}
\details{
Plot of adjusted Wald-type interval using error due to the
difference of achieved and nominal level of significance for the \eqn{n + 1} intervals
}
\examples{
n=20; alp=0.05; h=2;phi=0.99; f=-2
PloterrAWD(n,alp,h,phi,f)
}
\seealso{
Other Error for adjusted methods: 
\code{\link{PloterrAAS}()},
\code{\link{PloterrAAll}()},
\code{\link{PloterrALR}()},
\code{\link{PloterrALT}()},
\code{\link{PloterrASC}()},
\code{\link{PloterrATW}()},
\code{\link{errAAS}()},
\code{\link{errAAll}()},
\code{\link{errALR}()},
\code{\link{errALT}()},
\code{\link{errASC}()},
\code{\link{errATW}()},
\code{\link{errAWD}()}
}
\concept{Error for adjusted methods}
