% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/502.Error-Failure_LimitBased_BASE_All_Graph.R
\name{PloterrBA}
\alias{PloterrBA}
\title{Plots error, long term power and pass/fail criteria using Bayesian method}
\usage{
PloterrBA(n, alp, phi, f, a, b)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{phi}{- Null hypothesis value}

\item{f}{- Failure criterion}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}
}
\description{
Plots error, long term power and pass/fail criteria using Bayesian method
}
\details{
Plot of Bayesian Highest Probability Density
(HPD) and two tailed intervals using error due to the difference of achieved and
nominal level of significance for the \eqn{n + 1} intervals
for the Beta - Binomial conjugate prior model for the probability of success \code{p}
}
\examples{
n=20; alp=0.05; phi=0.05; f=-2;a=0.5;b=0.5
PloterrBA(n,alp,phi,f,a,b)
}
\seealso{
Other Error for base methods: \code{\link{PloterrAS}},
  \code{\link{PloterrAll}}, \code{\link{PloterrEX}},
  \code{\link{PloterrLR}}, \code{\link{PloterrLT}},
  \code{\link{PloterrSC}}, \code{\link{PloterrTW}},
  \code{\link{PloterrWD}}, \code{\link{errAS}},
  \code{\link{errAll}}, \code{\link{errBA}},
  \code{\link{errEX}}, \code{\link{errLR}},
  \code{\link{errLT}}, \code{\link{errSC}},
  \code{\link{errTW}}, \code{\link{errWD}}
}

