% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/522.Error-Failure_Limitbased_CC_All_Graph.R
\name{PloterrCAll}
\alias{PloterrCAll}
\title{Plots the calculated error, long term power and pass/fail criteria using 5 continuity
corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine)}
\usage{
PloterrCAll(n, alp, phi, c, f)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{phi}{- Null hypothesis value}

\item{c}{- Continuity correction}

\item{f}{- Failure criterion}
}
\description{
Plots the calculated error, long term power and pass/fail criteria using 5 continuity
corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine)
}
\details{
Plots the  error, long term power and pass/fail criteria calculaed using 5
continuity corrected methods (Wald, Wald-T, Score, Logit-Wald, ArcSine)
}
\examples{
n=5; alp=0.05; phi=0.05; c=1/(2*n); f=-2
PloterrCAll(n,alp,phi,c,f)
}
\seealso{
Other Error for continuity corrected methods: \code{\link{PloterrCAS}},
  \code{\link{PloterrCLT}}, \code{\link{PloterrCSC}},
  \code{\link{PloterrCTW}}, \code{\link{PloterrCWD}},
  \code{\link{errCAS}}, \code{\link{errCAll}},
  \code{\link{errCLT}}, \code{\link{errCSC}},
  \code{\link{errCTW}}, \code{\link{errCWD}}
}

