% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/522.Error-Failure_Limitbased_CC_All_Graph.R
\name{PloterrCWD}
\alias{PloterrCWD}
\title{Plots the error, long term power and pass/fail criteria for continuity corrected Wald method}
\usage{
PloterrCWD(n, alp, phi, c, f)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{phi}{- Null hypothesis value}

\item{c}{- Continuity correction}

\item{f}{- Failure criterion}
}
\description{
Plots the error, long term power and pass/fail criteria for continuity corrected Wald method
}
\details{
Plot of Wald-type interval with continuity correction using error
due to the difference of achieved and nominal level of significance for the \eqn{n + 1} intervals
}
\examples{
n=5; alp=0.05; phi=0.05;c=1/(2*n); f=-2
PloterrCWD(n,alp,phi,c,f)
}
\seealso{
Other Error for continuity corrected methods: 
\code{\link{PloterrCAS}()},
\code{\link{PloterrCAll}()},
\code{\link{PloterrCLT}()},
\code{\link{PloterrCSC}()},
\code{\link{PloterrCTW}()},
\code{\link{errCAS}()},
\code{\link{errCAll}()},
\code{\link{errCLT}()},
\code{\link{errCSC}()},
\code{\link{errCTW}()},
\code{\link{errCWD}()}
}
\concept{Error for continuity corrected methods}
