% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/302.Expec_Leng_BASE_All_Graph.R
\name{PlotexplAll}
\alias{PlotexplAll}
\title{Plots the Expected length using 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)}
\usage{
PlotexplAll(n, alp, a, b)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}
}
\description{
Plots the Expected length using 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\details{
The  plots using 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine) for the expected length of \code{n} given \code{alp}, \code{h}, \code{a}, \code{b}, \code{t1} and  \code{t2} using all the methods
}
\examples{
n= 10; alp=0.05; a=1;b=1;
PlotexplAll(n,alp,a,b)
}
\seealso{
Other Expected length  of base methods: \code{\link{PlotexplAS}},
  \code{\link{PlotexplBA}}, \code{\link{PlotexplEX}},
  \code{\link{PlotexplLR}}, \code{\link{PlotexplLT}},
  \code{\link{PlotexplSC}}, \code{\link{PlotexplTW}},
  \code{\link{PlotexplWD}}, \code{\link{PlotlengthAS}},
  \code{\link{PlotlengthAll}}, \code{\link{PlotlengthBA}},
  \code{\link{PlotlengthEX}}, \code{\link{PlotlengthLR}},
  \code{\link{PlotlengthLT}}, \code{\link{PlotlengthSC}},
  \code{\link{PlotlengthTW}}, \code{\link{PlotlengthWD}},
  \code{\link{lengthAS}}, \code{\link{lengthAll}},
  \code{\link{lengthBA}}, \code{\link{lengthEX}},
  \code{\link{lengthLR}}, \code{\link{lengthLT}},
  \code{\link{lengthSC}}, \code{\link{lengthTW}},
  \code{\link{lengthWD}}
}

