% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/302.Expec_Leng_BASE_All_Graph.R
\name{PlotexplAS}
\alias{PlotexplAS}
\title{Plots ArcSine method of expected length}
\usage{
PlotexplAS(n, alp, a, b)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}
}
\description{
Plots ArcSine method of expected length
}
\details{
Plot of Wald-type interval for the arcsine transformation of the parameter
\code{p} using sum of length of the \eqn{n + 1} intervals
}
\examples{
\dontrun{
n=5; alp=0.05;a=1;b=1
PlotexplAS(n,alp,a,b)
}
}
\seealso{
Other Expected length  of base methods: 
\code{\link{PlotexplAll}()},
\code{\link{PlotexplBA}()},
\code{\link{PlotexplEX}()},
\code{\link{PlotexplLR}()},
\code{\link{PlotexplLT}()},
\code{\link{PlotexplSC}()},
\code{\link{PlotexplTW}()},
\code{\link{PlotexplWD}()},
\code{\link{PlotlengthAS}()},
\code{\link{PlotlengthAll}()},
\code{\link{PlotlengthBA}()},
\code{\link{PlotlengthEX}()},
\code{\link{PlotlengthLR}()},
\code{\link{PlotlengthLT}()},
\code{\link{PlotlengthSC}()},
\code{\link{PlotlengthTW}()},
\code{\link{PlotlengthWD}()},
\code{\link{lengthAS}()},
\code{\link{lengthAll}()},
\code{\link{lengthBA}()},
\code{\link{lengthEX}()},
\code{\link{lengthLR}()},
\code{\link{lengthLT}()},
\code{\link{lengthSC}()},
\code{\link{lengthTW}()},
\code{\link{lengthWD}()}
}
\concept{Expected length  of base methods}
