% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/325.Expec_Leng_GENERAL_GIVENp.R
\name{PlotexplGEN}
\alias{PlotexplGEN}
\title{Plot the expected length given hypothetical "p"}
\usage{
PlotexplGEN(n, LL, UL, hp)
}
\arguments{
\item{n}{- Number of trials}

\item{LL}{- Lower limit}

\item{UL}{- Upper limit}

\item{hp}{- Hypothetical "p"}
}
\description{
Plot the expected length given hypothetical "p"
}
\details{
The  plot of the expected length for \code{n} given lower limit \code{LL} and  upper limit \code{UL}
}
\examples{
n= 5;
LL=c(0,0.01,0.0734,0.18237,0.3344,0.5492)		#Lower and Upper Limits
UL=c(0.4507,0.6655,0.8176,0.9265,0.9899,1)
hp=seq(0,1,by=0.01)
PlotexplGEN(n,LL,UL,hp)
}
\seealso{
Other Expected length: \code{\link{PlotexplSIM}},
  \code{\link{sumlenGEN}}
}

