% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/326.Expec_Leng_GENERAL_SIMULATEDp.R
\name{PlotexplSIM}
\alias{PlotexplSIM}
\title{Plots the expected length using calculated using simulation}
\usage{
PlotexplSIM(n, LL, UL, s, a, b)
}
\arguments{
\item{n}{- Number of trials}

\item{LL}{- Lower limit}

\item{UL}{- Upper limit}

\item{s}{- Number of Hypothetical "p"}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}
}
\description{
Plots the expected length using calculated using simulation
}
\details{
The plot of the expected length for \code{n} given lower limit \code{LL} and  upper limit \code{UL}
}
\examples{
LL=c(0,0.01,0.0734,0.18237,0.3344,0.5492)		#Lower and Upper Limits
UL=c(0.4507,0.6655,0.8176,0.9265,0.9899,1)
n= 5; s=5000; a=1; b=1;
PlotexplSIM(n,LL,UL,s,a,b)
}
\seealso{
Other Expected length: \code{\link{PlotexplGEN}},
  \code{\link{PlotlengthGEN}}, \code{\link{PlotlengthSIM}},
  \code{\link{lengthGEN}}, \code{\link{lengthSIM}}
}

