% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/303.Sum_Leng_BASE_All_Graph.R
\name{PlotlengthAll}
\alias{PlotlengthAll}
\title{Plots the length summary for 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)}
\usage{
PlotlengthAll(n, alp, a, b)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}
}
\description{
Plots the length summary for 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\details{
The  plots for 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine).
Sum of the length is shown as bar graph.
}
\examples{
\dontrun{
n= 10; alp=0.05; a=1;b=1;
PlotlengthAll(n,alp,a,b)
}
}
\seealso{
Other Expected length  of base methods: 
\code{\link{PlotexplAS}()},
\code{\link{PlotexplAll}()},
\code{\link{PlotexplBA}()},
\code{\link{PlotexplEX}()},
\code{\link{PlotexplLR}()},
\code{\link{PlotexplLT}()},
\code{\link{PlotexplSC}()},
\code{\link{PlotexplTW}()},
\code{\link{PlotexplWD}()},
\code{\link{PlotlengthAS}()},
\code{\link{PlotlengthBA}()},
\code{\link{PlotlengthEX}()},
\code{\link{PlotlengthLR}()},
\code{\link{PlotlengthLT}()},
\code{\link{PlotlengthSC}()},
\code{\link{PlotlengthTW}()},
\code{\link{PlotlengthWD}()},
\code{\link{lengthAS}()},
\code{\link{lengthAll}()},
\code{\link{lengthBA}()},
\code{\link{lengthEX}()},
\code{\link{lengthLR}()},
\code{\link{lengthLT}()},
\code{\link{lengthSC}()},
\code{\link{lengthTW}()},
\code{\link{lengthWD}()}
}
\concept{Expected length  of base methods}
