% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/323.Sum_Leng_CC_All_Graph.R
\name{PlotlengthCWD}
\alias{PlotlengthCWD}
\title{Plots the sum of length using continuity corrected Wald method}
\usage{
PlotlengthCWD(n, alp, c, a, b)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{c}{- Continuity correctection}

\item{a}{- Beta parameters for hypo "p"}

\item{b}{- Beta parameters for hypo "p"}
}
\description{
Plots the sum of length using continuity corrected Wald method
}
\details{
The  plots of the expected length of continuity corrected Wald method
}
\examples{
n= 10; alp=0.05; c=1/(2*n);a=1;b=1;
PlotlengthCWD(n,alp,c,a,b)
}
\seealso{
Other Expected length  of continuity corrected methods: \code{\link{PlotexplCAS}},
  \code{\link{PlotexplCAll}}, \code{\link{PlotexplCLT}},
  \code{\link{PlotexplCSC}}, \code{\link{PlotexplCTW}},
  \code{\link{PlotexplCWD}}, \code{\link{PlotlengthCAS}},
  \code{\link{PlotlengthCAll}},
  \code{\link{PlotlengthCLT}}, \code{\link{PlotlengthCSC}},
  \code{\link{PlotlengthCTW}}, \code{\link{lengthCAS}},
  \code{\link{lengthCAll}}, \code{\link{lengthCLT}},
  \code{\link{lengthCSC}}, \code{\link{lengthCTW}},
  \code{\link{lengthCWD}}
}

