% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/412.p-Confidence_p-Bias_ADJ_All_Graph.R
\name{PlotpCOpBIAAll}
\alias{PlotpCOpBIAAll}
\title{Plots  p-confidence and p-bias for a given n and alpha level for 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)}
\usage{
PlotpCOpBIAAll(n, alp, h)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\description{
Plots  p-confidence and p-bias for a given n and alpha level for 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\details{
Plots of p-confidence and p-bias plots for 6 adjusted methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\examples{
n=5; alp=0.05;h=2
PlotpCOpBIAAll(n,alp,h)
}
\seealso{
Other p-confidence and p-bias of adjusted methods: \code{\link{PlotpCOpBIAAS}},
  \code{\link{PlotpCOpBIALR}}, \code{\link{PlotpCOpBIALT}},
  \code{\link{PlotpCOpBIASC}}, \code{\link{PlotpCOpBIATW}},
  \code{\link{PlotpCOpBIAWD}}, \code{\link{pCOpBIAAS}},
  \code{\link{pCOpBIAAll}}, \code{\link{pCOpBIALR}},
  \code{\link{pCOpBIALT}}, \code{\link{pCOpBIASC}},
  \code{\link{pCOpBIATW}}, \code{\link{pCOpBIAWD}}
}

