% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/402.p-Confidence_p-Bias_BASE_All_Graph.R
\name{PlotpCOpBIAll}
\alias{PlotpCOpBIAll}
\title{Plots  p-confidence and p-bias for a given n and alpha level for
6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)}
\usage{
PlotpCOpBIAll(n, alp)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}
}
\description{
Plots  p-confidence and p-bias for a given n and alpha level for
6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\details{
p-confidence and p-bias plots for 6 base methods (Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
}
\examples{
\dontrun{
n=5; alp=0.05
PlotpCOpBIAll(n,alp)
}
}
\references{
[1] 2005 Vos PW and Hudson S.
Evaluation Criteria for Discrete Confidence Intervals: Beyond Coverage and Length.
The American Statistician: 59; 137 - 142.
}
\seealso{
Other p-confidence and p-bias of base methods: \code{\link{PlotpCOpBIAS}},
  \code{\link{PlotpCOpBIBA}}, \code{\link{PlotpCOpBIEX}},
  \code{\link{PlotpCOpBILR}}, \code{\link{PlotpCOpBILT}},
  \code{\link{PlotpCOpBISC}}, \code{\link{PlotpCOpBITW}},
  \code{\link{PlotpCOpBIWD}}, \code{\link{pCOpBIAS}},
  \code{\link{pCOpBIAll}}, \code{\link{pCOpBIBA}},
  \code{\link{pCOpBIEX}}, \code{\link{pCOpBILR}},
  \code{\link{pCOpBILT}}, \code{\link{pCOpBISC}},
  \code{\link{pCOpBITW}}, \code{\link{pCOpBIWD}}
}

