% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/412.p-Confidence_p-Bias_ADJ_All_Graph.R
\name{PlotpCOpBIASC}
\alias{PlotpCOpBIASC}
\title{Plots  p-confidence and p-bias for adjusted Score method}
\usage{
PlotpCOpBIASC(n, alp, h)
}
\arguments{
\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\description{
Plots  p-confidence and p-bias for adjusted Score method
}
\details{
p-confidence and p-bias plots for adjusted Score method
}
\examples{
n=5; alp=0.05;h=2
PlotpCOpBIASC(n,alp,h)
}
\seealso{
Other p-confidence and p-bias of adjusted methods: \code{\link{PlotpCOpBIAAS}},
  \code{\link{PlotpCOpBIAAll}},
  \code{\link{PlotpCOpBIALR}}, \code{\link{PlotpCOpBIALT}},
  \code{\link{PlotpCOpBIATW}}, \code{\link{PlotpCOpBIAWD}},
  \code{\link{pCOpBIAAS}}, \code{\link{pCOpBIAAll}},
  \code{\link{pCOpBIALR}}, \code{\link{pCOpBIALT}},
  \code{\link{pCOpBIASC}}, \code{\link{pCOpBIATW}},
  \code{\link{pCOpBIAWD}}
}

