% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/423.p-Confidence_p-Bias_GENERAL.R
\name{PlotpCOpBIGEN}
\alias{PlotpCOpBIGEN}
\title{Plots p-Confidence and p-Bias estimation for general method}
\usage{
PlotpCOpBIGEN(n, LL, UL)
}
\arguments{
\item{n}{- Number of trials}

\item{LL}{- Lower limit}

\item{UL}{- Upper limit}
}
\description{
Plots p-Confidence and p-Bias estimation for general method
}
\details{
Plot of the general method - intervals obtained from any method using
p-confidence and p-bias for the  \eqn{n + 1} intervals
}
\examples{
LL=c(0,0.01,0.0734,0.18237,0.3344,0.5492)		#Lower and Upper Limits
UL=c(0.4507,0.6655,0.8176,0.9265,0.9899,1)
n=5;
PlotpCOpBIGEN(n,LL,UL)
}
\seealso{
Other General methods for p-Confidence and p-Bias: 
\code{\link{pCOpBIGEN}()}
}
\concept{General methods for p-Confidence and p-Bias}
