% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/621.Poster_Predictive.R
\name{probPRE}
\alias{probPRE}
\title{The Predicted probability - Bayesian approach}
\usage{
probPRE(n, m, a1, a2)
}
\arguments{
\item{n}{- Number of trials from data}

\item{m}{- Future :Number of trials}

\item{a1}{- Beta Prior Parameters for Bayesian estimation}

\item{a2}{- Beta Prior Parameters for Bayesian estimation}
}
\value{
A matrix of probability values between [0,1]
 \item{predicted_probability }{- The predicted probability}
 \item{0:n}{The number of columns based on the value of n}
}
\description{
The Predicted probability - Bayesian approach
}
\details{
Computes posterior predictive probabilities for the required size of number of
trials \code{m} from the given number of trials \code{n} for the given parameters for Beta prior
distribution
}
\examples{
n=10; m=5; a1=0.5; a2=0.5
probPRE(n,m,a1,a2)
}
\references{
[1] 2002 Gelman A, Carlin  JB, Stern HS and Dunson DB
Bayesian Data Analysis, Chapman & Hall/CRC
}
\seealso{
Other Miscellaneous  functions for Bayesian method: 
\code{\link{empericalBAx}()},
\code{\link{empericalBA}()},
\code{\link{probPOSx}()},
\code{\link{probPOS}()},
\code{\link{probPREx}()}
}
\concept{Miscellaneous  functions for Bayesian method}
