% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/113.ConfidenceIntervals_ADJ_n_x.R
\name{PlotciAAllx}
\alias{PlotciAAllx}
\title{Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score,
Logit-Wald, ArcSine)}
\usage{
PlotciAAllx(x, n, alp, h)
}
\arguments{
\item{x}{- Number of success}

\item{n}{- Number of trials}

\item{alp}{- Alpha value (significance level required)}

\item{h}{- Adding factor}
}
\description{
Plots the CI estimation of 6 adjusted methods (Wald, Wald-T, Likelihood, Score,
Logit-Wald, ArcSine)
}
\details{
The plots of confidence intervals of using 6 adjusted methods
(Wald, Wald-T, Likelihood, Score, Logit-Wald, ArcSine)
for \code{n} given \code{alp}, \code{x} and \code{h}
}
\examples{
x=5; n=5; alp=0.05; h=2
PlotciAAllx(x,n,alp,h)
}
\seealso{
Other Adjusted methods of CI estimation  given x & n: 
\code{\link{ciAASx}()},
\code{\link{ciAAllx}()},
\code{\link{ciALRx}()},
\code{\link{ciALTx}()},
\code{\link{ciASCx}()},
\code{\link{ciATWx}()},
\code{\link{ciAWDx}()}
}
\concept{Adjusted methods of CI estimation  given x & n}
